+++ NEWS.md
@@ -1,3 +1,7 @@
+# dev
+
+* Added 'variable' argument to 'nc_atts', per #8. 
+
 # ncmeta 0.0.2
 
 * added some extra checks for bad source strings, particularly the empty string